test(dialog): use dialog helper in tests
