+++ packages/html/src/dialog/tests/dialog-action-buttons.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en" class="k-no-animations">
 <head>
-    <title>Dialog Action Buttons</title>
+    <title>KUI React Dialog Action Buttons</title>
     <meta charset="utf-8" />
     <link rel="stylesheet" data-role="kendo-theme" href="/packages/default/dist/all.css"/>
     <link rel="stylesheet" href="/packages/html/assets/styles.css"/>
