+++ packages/html/src/dialog/tests/dialog-in-window.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { Button } from '../../button';
+import { Window } from '../../window';
+import { Dialog } from '../../dialog';
 
 const root = ReactDOM.createRoot(
     document.getElementById('app') as HTMLElement
@@ -27,33 +28,16 @@ root.render(
         <style>{styles}</style>
         <div id="test-area" className="k-d-grid k-grid-cols-1">
             <section>
-                <div className="k-widget k-window k-window-wrapper">
-                    <div className="k-window-titlebar k-header">
-                        <div className="k-window-title">Window with dialog</div>
-                        <div className="k-window-actions">
-                            <Button fillMode="flat" icon="window-minimize"></Button>
-                            <Button fillMode="flat" icon="window-maximize"></Button>
-                            <Button fillMode="flat" icon="close"></Button>
-                        </div>
+
+                <Window title="Window with dialog" actions={[ 'window-minimize', 'window-maximize', 'close' ]} className="k-window-wrapper">
+                    Window content
+                    <div className="k-dialog-wrapper">
+                        <div className="k-overlay"></div>
+                        <Dialog title="Dialog in window" actions={[ 'close' ]}>
+                            Dialog content
+                        </Dialog>
                     </div>
-                    <div className="k-window-content">
-                        Window content
-                        <div className="k-dialog-wrapper">
-                            <div className="k-overlay"></div>
-                            <div className="k-widget k-window k-dialog">
-                                <div className="k-window-titlebar k-dialog-titlebar k-header">
-                                    <div className="k-window-title k-dialog-title">Dialog in window</div>
-                                    <div className="k-window-actions k-dialog-actions">
-                                        <Button fillMode="flat" icon="close"></Button>
-                                    </div>
-                                </div>
-                                <div className="k-window-content k-dialog-content">
-                                    Dialog content
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
+                </Window>
             </section>
         </div>
     </>
