+++ packages/html/src/dialog/tests/dialog-modal.html
@@ -1,13 +1,13 @@
 <!DOCTYPE html>
 <html lang="en" class="k-no-animations">
 <head>
-    <title>KUI React Dialog Action Buttons</title>
+    <title>KUI Angular Dialog Modal</title>
     <meta charset="utf-8" />
     <link rel="stylesheet" data-role="kendo-theme" href="/packages/default/dist/all.css"/>
     <link rel="stylesheet" href="/packages/html/assets/styles.css"/>
     <script src="/packages/html/assets/scripts.js"></script>
 </head>
 <body id="app" class="k-body">
-    <script src="/packages/html/dist/dialog/tests/dialog-action-buttons-react.js"></script>
+    <script src="/packages/html/dist/dialog/tests/dialog-modal.js"></script>
 </body>
 </html>
