+++ packages/html/src/dialog/tests/dialog.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { ActionButtons } from '../../action-buttons';
-import { Button } from '../../button';
+import { Dialog } from '../../dialog';
+
 
 const root = ReactDOM.createRoot(
     document.getElementById('app') as HTMLElement
@@ -18,105 +18,35 @@ root.render(
         <style>{styles}</style>
         <div id="test-area" className="k-d-grid k-grid-cols-3">
 
-            <div className="k-widget k-window k-dialog">
-                <div className="k-window-titlebar k-dialog-titlebar">
-                    <span className="k-window-title k-dialog-title">
-                        Title
+            <Dialog title="Title
                         g y p
-                        t d l
-                    </span>
-                    <div className="k-window-actions k-dialog-actions">
-                        <Button fillMode="flat" icon="close"></Button>
-                    </div>
-                </div>
+                        t d l" actions={[ 'close' ]} actionButtons actionButtonsAlign="end">
+                Dialog content template
+            </Dialog>
 
-                <div className="k-window-content k-dialog-content">
-                    Dialog content template
-                </div>
-
-                <ActionButtons className="k-dialog-buttongroup" alignment="end">
-                    <Button>Action</Button>
-                    <Button themeColor="primary">Primary</Button>
-                </ActionButtons>
-            </div>
-
-            <div className="k-widget k-window k-dialog">
-                <div className="k-window-titlebar k-dialog-titlebar">
-                    <span className="k-window-title k-dialog-title" style={{ fontFamily: "cursive" }}>
-                        Cursive
+            <Dialog title="Cursive
                         g y p
-                        t d l
-                    </span>
-                </div>
+                        t d l" actionButtons actionButtonsAlign="end">
+                Dialog without titlebar actions
+            </Dialog>
 
-                <div className="k-window-content k-dialog-content">
-                    Dialog without titlebar actions
-                </div>
+            <Dialog title="" actions={[ 'close' ]} actionButtons actionButtonsAlign="end">
+                Dialog with empty title and actions
+            </Dialog>
 
-                <ActionButtons className="k-dialog-buttongroup" alignment="end">
-                    <Button>Action</Button>
-                    <Button themeColor="primary">Primary</Button>
-                </ActionButtons>
-            </div>
+            <Dialog title="">
+                Dialog with empty title and no actions
+            </Dialog>
 
-            <div className="k-widget k-window k-dialog">
-                <div className="k-window-titlebar k-dialog-titlebar">
-                    <span className="k-window-title k-dialog-title"></span>
-                    <div className="k-window-actions k-dialog-actions">
-                        <Button fillMode="flat" icon="close"></Button>
-                    </div>
-                </div>
+            <Dialog actionButtons actionButtonsAlign="end">
+                Dialog without title
+            </Dialog>
 
-                <div className="k-window-content k-dialog-content">
-                    Dialog with empty title and actions
-                </div>
-
-                <ActionButtons className="k-dialog-buttongroup" alignment="end">
-                    <Button>Action</Button>
-                    <Button themeColor="primary">Primary</Button>
-                </ActionButtons>
-            </div>
-
-            <div className="k-widget k-window k-dialog">
-                <div className="k-window-titlebar k-dialog-titlebar">
-                    <span className="k-window-title k-dialog-title"></span>
-                </div>
-
-                <div className="k-window-content k-dialog-content">
-                    Dialog with empty title and no actions
-                </div>
-            </div>
-
-            <div className="k-widget k-window k-dialog">
-                <div className="k-window-content k-dialog-content">
-                    Dialog without title
-                </div>
-
-                <ActionButtons className="k-dialog-buttongroup" alignment="end">
-                    <Button>Action</Button>
-                    <Button themeColor="primary">Primary</Button>
-                </ActionButtons>
-            </div>
-
-            <div className="k-widget k-window k-dialog" style={{ height: "170px" }}>
-                <div className="k-window-titlebar k-dialog-titlebar">
-                    <span className="k-window-title k-dialog-title">Scrolling</span>
-                    <div className="k-window-actions k-dialog-actions">
-                        <Button fillMode="flat" icon="close"></Button>
-                    </div>
-                </div>
-
-                <div className="k-window-content k-dialog-content k-scroll">
-                    <p>Dialog content template with long text and vertical scrolling.</p>
-                    <p>Dialog content template with long text and vertical scrolling.</p>
-                    <p>Dialog content template with long text and vertical scrolling.</p>
-                </div>
-
-                <ActionButtons className="k-dialog-buttongroup" alignment="end">
-                    <Button>Action</Button>
-                    <Button themeColor="primary">Primary</Button>
-                </ActionButtons>
-            </div>
+            <Dialog title="Scrolling" actions={[ 'close' ]} actionButtons actionButtonsAlign="end" style={{ height: "170px" }}>
+                <p>Dialog content template with long text and vertical scrolling.</p>
+                <p>Dialog content template with long text and vertical scrolling.</p>
+                <p>Dialog content template with long text and vertical scrolling.</p>
+            </Dialog>
 
         </div>
     </>
