--- tests/dialog/dialog-action-buttons-react.html
@@ -1,97 +1,0 @@
-<!doctype html>
-<html lang="en" class="k-no-animations" style="--kendo-scrollbar-width:15px;">
-<head>
-    <title>KUI React Dialog Action Buttons</title>
-    <meta charset="utf-8">
-    <link rel="stylesheet" data-role="kendo-theme" href="/packages/default/dist/all.css">
-    <link rel="stylesheet" href="/packages/html/assets/styles.css">
-    <script src="/packages/html/assets/scripts.js"></script>
-</head>
-<body id="app" class="k-body">
-    <style>
-        .k-dialog {
-            position: relative;
-            width: 300px;
-        }
-
-    </style>
-    <div id="test-area" class="k-d-grid k-grid-cols-2">
-        <div class="k-widget k-window k-dialog">
-            <div class="k-window-titlebar k-dialog-titlebar k-header">
-                <div class="k-window-title k-dialog-title">Title</div>
-                <div class="k-window-actions k-dialog-actions">
-                    <button class="k-button k-button-md k-button-flat k-button-flat-base k-rounded-md k-icon-button">
-                        <span class="k-button-icon k-icon k-i-close"></span>
-                    </button>
-                </div>
-            </div>
-            <div class="k-window-content k-dialog-content">Action Buttons at start position</div>
-            <div class="k-actions k-actions-start k-actions-horizontal k-dialog-buttongroup">
-                <button class="k-button k-button-md k-button-solid k-button-solid-base k-rounded-md">
-                    <span class="k-button-text">Action</span>
-                </button>
-                <button class="k-button k-button-md k-button-solid k-button-solid-primary k-rounded-md">
-                    <span class="k-button-text">Primary</span>
-                </button>
-            </div>
-        </div>
-        <div class="k-widget k-window k-dialog">
-            <div class="k-window-titlebar k-dialog-titlebar k-header">
-                <div class="k-window-title k-dialog-title">Title</div>
-                <div class="k-window-actions k-dialog-actions">
-                    <button class="k-button k-button-md k-button-flat k-button-flat-base k-rounded-md k-icon-button">
-                        <span class="k-button-icon k-icon k-i-close"></span>
-                    </button>
-                </div>
-            </div>
-            <div class="k-window-content k-dialog-content">Action Buttons at center position</div>
-            <div class="k-actions k-actions-center k-actions-horizontal k-dialog-buttongroup">
-                <button class="k-button k-button-md k-button-solid k-button-solid-base k-rounded-md">
-                    <span class="k-button-text">Action</span>
-                </button>
-                <button class="k-button k-button-md k-button-solid k-button-solid-primary k-rounded-md">
-                    <span class="k-button-text">Primary</span>
-                </button>
-            </div>
-        </div>
-        <div class="k-widget k-window k-dialog">
-            <div class="k-window-titlebar k-dialog-titlebar k-header">
-                <div class="k-window-title k-dialog-title">Title</div>
-                <div class="k-window-actions k-dialog-actions">
-                    <button class="k-button k-button-md k-button-flat k-button-flat-base k-rounded-md k-icon-button">
-                        <span class="k-button-icon k-icon k-i-close"></span>
-                    </button>
-                </div>
-            </div>
-            <div class="k-window-content k-dialog-content">Action Buttons at end position</div>
-            <div class="k-actions k-actions-end k-actions-horizontal k-dialog-buttongroup">
-                <button class="k-button k-button-md k-button-solid k-button-solid-base k-rounded-md">
-                    <span class="k-button-text">Action</span>
-                </button>
-                <button class="k-button k-button-md k-button-solid k-button-solid-primary k-rounded-md">
-                    <span class="k-button-text">Primary</span>
-                </button>
-            </div>
-        </div>
-        <div class="k-widget k-window k-dialog">
-            <div class="k-window-titlebar k-dialog-titlebar k-header">
-                <div class="k-window-title k-dialog-title">Title</div>
-                <div class="k-window-actions k-dialog-actions">
-                    <button class="k-button k-button-md k-button-flat k-button-flat-base k-rounded-md k-icon-button">
-                        <span class="k-button-icon k-icon k-i-close"></span>
-                    </button>
-                </div>
-            </div>
-            <div class="k-window-content k-dialog-content">Action Buttons with stretched position</div>
-            <div class="k-actions k-actions-stretched k-actions-horizontal k-dialog-buttongroup">
-                <button class="k-button k-button-md k-button-solid k-button-solid-base k-rounded-md">
-                    <span class="k-button-text">Action</span>
-                </button>
-                <button class="k-button k-button-md k-button-solid k-button-solid-primary k-rounded-md">
-                    <span class="k-button-text">Primary</span>
-                </button>
-            </div>
-        </div>
-    </div>
-</body>
-</html>
