--- tests/dialog/dialog-modal-angular.html
@@ -1,48 +1,0 @@
-<!doctype html>
-<html lang="en" class="k-no-animations" style="--kendo-scrollbar-width:15px;">
-<head>
-    <title>KUI Angular Dialog Modal</title>
-    <meta charset="utf-8">
-    <link rel="stylesheet" data-role="kendo-theme" href="/packages/default/dist/all.css">
-    <link rel="stylesheet" href="/packages/html/assets/styles.css">
-    <script src="/packages/html/assets/scripts.js"></script>
-</head>
-<body id="app" class="k-body">
-    <style>
-        #test-area {
-            width: 500px;
-            height: 400px;
-            transform: translate(0, 0);
-        }
-        .k-dialog {
-            width: 300px;
-            height: 200px;
-        }
-
-    </style>
-    <div id="test-area" class="k-grid-cols-1">
-        <div class="k-dialog-wrapper">
-            <div class="k-overlay"></div>
-            <div class="k-widget k-window k-dialog">
-                <div class="k-window-titlebar k-dialog-titlebar k-header">
-                    <div class="k-window-title k-dialog-title">Title</div>
-                    <div class="k-window-actions k-dialog-actions">
-                        <button class="k-button k-button-md k-button-flat k-button-flat-base k-rounded-md k-icon-button">
-                            <span class="k-button-icon k-icon k-i-close"></span>
-                        </button>
-                    </div>
-                </div>
-                <div class="k-window-content k-dialog-content">Dialog content template</div>
-                <div class="k-actions k-actions-end k-actions-horizontal k-dialog-buttongroup">
-                    <button class="k-button k-button-md k-button-solid k-button-solid-base k-rounded-md">
-                        <span class="k-button-text">Action</span>
-                    </button>
-                    <button class="k-button k-button-md k-button-solid k-button-solid-primary k-rounded-md">
-                        <span class="k-button-text">Primary</span>
-                    </button>
-                </div>
-            </div>
-        </div>
-    </div>
-</body>
-</html>
